+++ website/src/examples/react/image-view/sample-doc-image.ts
