+++ website/src/examples/vue/image-view/editor.vue
@@ -9,8 +9,8 @@ import {
   watchPostEffect,
 } from 'vue'
 
-import { defaultContent } from './default-doc'
 import { defineExtension } from './extension'
+import { defaultContent } from './sample-doc-image'
 
 const editor = createEditor({ extension: defineExtension(), defaultContent })
 const editorRef = ref<HTMLDivElement | null>(null)
