+++ website/src/examples/vue/image-view/sample-doc-image.ts
